Refactor e2e - part 5
